tools-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
httpx==0.24.0
jsonschema==4.17.3
pymongo==4.3.3
aiohttp==3.8.4
redis==4.5.4
prometheus-client==0.16.0
pytest==7.3.1
requests==2.28.2
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
pytz==2023.3
